Refactor Neon.js WIP
