+++ doc/Color.js.html
@@ -1,0 +1,121 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: Color.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: Color.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>/** @module Color*/
+
+/**
+ * Highlight each staff a different color.
+ */
+export function setStaffHighlight() {
+    let staves = Array.from(document.getElementsByClassName("staff"));
+    for (var i = 0; i &lt; staves.length; i++) {
+        let staffColor = ColorPalette[i % ColorPalette.length];
+        highlight(staves[i], staffColor);
+    }
+}
+
+/**
+ * Remove the highlight from each staff.
+ */
+export function unsetStaffHighlight() {
+    unhighlight(".staff");
+}
+
+/**
+ * Highlight a staff a certain color.
+ * @param {SVGSVGElement} staff - The staff's SVG element.
+ * @param {string} color - The color to highlight the staff.
+ */
+export function highlight(staff, color) {
+    let children = Array.from($("#" + staff.id).children());
+    children.forEach(child => {
+        if (child.tagName === "path") {
+            child.setAttribute("stroke", color);
+        } else {
+            child.setAttribute("fill", color);
+        }
+        let currentClass = child.getAttribute("class");
+        child.setAttribute("class", currentClass !== null ? currentClass + " highlighted" : "highlighted");
+    });
+}
+
+/**
+ * Remove the highlight from a staff.
+ * @param {(SVGSVGElement|string)} staff - The staff's SVG element or a JQuery selector.
+ */
+export function unhighlight(staff) {
+    let children = Array.from($(staff).children(".highlighted"));
+    children.forEach(elem => {
+        if (elem.tagName === "path") {
+            elem.setAttribute("stroke", "#000000");
+        } else {
+            elem.removeAttribute("fill");
+        }
+    });
+    $(staff).children(".highlighted").removeClass("highlighted");
+}
+    
+/**
+ * Color palette from Figure 2 (Colors optimized for color-blind
+ * individuals) from "Points of view: Color blindness" by Bang Wong
+ * published in Nature Methods volume 8 on 27 May 2011
+ * https://www.nature.com/articles/nmeth.1618?WT.ec_id=NMETH-201106
+ */
+const ColorPalette = [
+    "rgb(0,0,0)",
+    "rgb(230, 159, 0)",
+    "rgb(86, 180, 233)",
+    "rgb(0, 158, 115)",
+    "rgb(240, 228, 66)",
+    "rgb(0, 114, 178)",
+    "rgb(213, 94, 0)",
+    "rgb(204, 121, 167)"
+];
+</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 10:56:27 GMT-0400 (EDT)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
