+++ doc/DragSelect.js.html
@@ -26,13 +26,13 @@
     
     <section>
         <article>
-            <pre class="prettyprint source linenums"><code>import ColorStaves, {highlight, unhighlight} from "./ColorStaves.js";
+            <pre class="prettyprint source linenums"><code>import * as Color from "./Color.js";
 
 /**
  * Handle dragging to select musical elements and staves.
  * @constructor
  * @param {DragHandler} dragHandler - Instantiated DragHandler object.
- * @param {module:ZoomHandler~ZoomHandler} zoomHandler - Instantiated ZoomHandler object.
+ * @param {module:Zoom~ZoomHandler} zoomHandler - Instantiated ZoomHandler object.
  * @param {GroupingHandler} groupingHandler - Instantiated GroupingHandler object.
  */
 function DragSelect (dragHandler, zoomHandler, groupingHandler) {
@@ -140,7 +140,7 @@ function DragSelect (dragHandler, zoomHandler, groupingHandler) {
                     }
                 });
                 toSelect.forEach(elem => {
-                    highlight(elem, "#d00");
+                    Color.highlight(elem, "#d00");
                     $(elem).addClass("selected");
                 });
             }
@@ -221,7 +221,7 @@ function DragSelect (dragHandler, zoomHandler, groupingHandler) {
         for (var i=0; i &lt; selected.length; i++) {
             if ($(selected[i]).hasClass("staff")) {
                 $(selected[i]).removeClass("selected");
-                unhighlight(selected[i]);
+                Color.unhighlight(selected[i]);
             } else {
                 $(selected[i]).removeClass("selected").attr("fill", null);
             }
@@ -230,8 +230,7 @@ function DragSelect (dragHandler, zoomHandler, groupingHandler) {
             groupingHandler.endGroupingSelection();
         }
         if ($("#highlightStaves").is(":checked")) {
-            let color = new ColorStaves();
-            color.setColor();
+            Color.setStaffHighlight();
         }
     }
 }
@@ -247,13 +246,13 @@ export {DragSelect as default};
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.html">ZoomHandler</a></li></ul><h3>Classes</h3><ul><li><a href="CursorHandler.html">CursorHandler</a></li><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertControls.html">InsertControls</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-ColorStaves-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.ViewBox.html">ViewBox</a></li><li><a href="module-ZoomHandler-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li><li><a href="ViewControls.html">ViewControls</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 08:46:07 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 10:56:27 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
