+++ doc/NeonView.html
@@ -404,7 +404,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line171">line 171</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line170">line 170</a>
     </li></ul></dd>
     
 
@@ -512,7 +512,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line109">line 109</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line108">line 108</a>
     </li></ul></dd>
     
 
@@ -598,7 +598,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line45">line 45</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line44">line 44</a>
     </li></ul></dd>
     
 
@@ -684,7 +684,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line187">line 187</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line186">line 186</a>
     </li></ul></dd>
     
 
@@ -788,7 +788,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line81">line 81</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line80">line 80</a>
     </li></ul></dd>
     
 
@@ -874,7 +874,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line118">line 118</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line117">line 117</a>
     </li></ul></dd>
     
 
@@ -960,7 +960,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line161">line 161</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line160">line 160</a>
     </li></ul></dd>
     
 
@@ -1046,7 +1046,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line92">line 92</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line91">line 91</a>
     </li></ul></dd>
     
 
@@ -1132,7 +1132,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line179">line 179</a>
+        <a href="NeonView.js.html">NeonView.js</a>, <a href="NeonView.js.html#line178">line 178</a>
     </li></ul></dd>
     
 
@@ -1194,13 +1194,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.html">ZoomHandler</a></li></ul><h3>Classes</h3><ul><li><a href="CursorHandler.html">CursorHandler</a></li><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertControls.html">InsertControls</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-ColorStaves-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.ViewBox.html">ViewBox</a></li><li><a href="module-ZoomHandler-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li><li><a href="ViewControls.html">ViewControls</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 08:46:07 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 10:56:27 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
