+++ doc/NeonView.js.html
@@ -27,9 +27,9 @@
     <section>
         <article>
             <pre class="prettyprint source linenums"><code>import Neon from "./Neon.js";
-import ZoomHandler from "./ZoomHandler.js";
+import ZoomHandler from "./Zoom.js";
 import InfoBox from "./InfoBox.js";
-import ViewControls from "./ViewControls.js";
+import * as Controls from "./Controls.js";
 import EditMode from "./EditMode.js";
 
 const verovio = require("verovio-dev");
@@ -52,19 +52,18 @@ function NeonView (params) {
     var neon = null;
     var zoomHandler = null;
     var infoBox = null;
-    var viewControls = null;
     var editMode = null;
 
     $.get(meiFile, (data) => {
         neon = new Neon(data, vrvToolkit);
         zoomHandler = new ZoomHandler();
         infoBox = new InfoBox(neon);
-        viewControls = new ViewControls(zoomHandler);
+        Controls.initDisplayControls(zoomHandler);
         editMode = new EditMode(this, meiFile, zoomHandler);
         loadView();
         // editMode.getScale();
-        viewControls.setSylControls();
-        viewControls.setHighlightControls();
+        Controls.setSylControls();
+        Controls.setHighlightControls();
     });
 
     /**
@@ -98,8 +97,8 @@ function NeonView (params) {
         else {
             loadSvg(); 
         }
-        viewControls.updateSylVisibility();
-        viewControls.updateHighlight();
+        Controls.updateSylVisibility();
+        Controls.updateHighlight();
         resetListeners();
     }
 
@@ -180,7 +179,7 @@ function NeonView (params) {
 
     function resetTransformations () {
         zoomHandler.restoreTransformation();
-        viewControls.setOpacityFromSlider();
+        Controls.setOpacityFromSlider();
     }
 
     /**
@@ -242,13 +241,13 @@ export {NeonView as default};
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.html">ZoomHandler</a></li></ul><h3>Classes</h3><ul><li><a href="CursorHandler.html">CursorHandler</a></li><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertControls.html">InsertControls</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-ColorStaves-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.ViewBox.html">ViewBox</a></li><li><a href="module-ZoomHandler-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li><li><a href="ViewControls.html">ViewControls</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 08:46:07 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 10:56:27 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
