+++ doc/ViewControls.html
@@ -91,7 +91,7 @@
             <td class="type">
             
                 
-<span class="param-type"><a href="module-ZoomHandler-ZoomHandler.html">module:ZoomHandler~ZoomHandler</a></span>
+<span class="param-type"><a href="module-Zoom-ZoomHandler.html">module:Zoom~ZoomHandler</a></span>
 
 
             
@@ -248,7 +248,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line60">line 60</a>
+        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line59">line 59</a>
     </li></ul></dd>
     
 
@@ -334,7 +334,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line89">line 89</a>
+        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line88">line 88</a>
     </li></ul></dd>
     
 
@@ -420,7 +420,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line45">line 45</a>
+        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line44">line 44</a>
     </li></ul></dd>
     
 
@@ -506,7 +506,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line76">line 76</a>
+        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line75">line 75</a>
     </li></ul></dd>
     
 
@@ -592,7 +592,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line28">line 28</a>
+        <a href="ViewControls.js.html">ViewControls.js</a>, <a href="ViewControls.js.html#line27">line 27</a>
     </li></ul></dd>
     
 
@@ -636,13 +636,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.html">ZoomHandler</a></li></ul><h3>Classes</h3><ul><li><a href="CursorHandler.html">CursorHandler</a></li><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertControls.html">InsertControls</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-ColorStaves-ColorStaves.html">ColorStaves</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.ViewBox.html">ViewBox</a></li><li><a href="module-ZoomHandler-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li><li><a href="ViewControls.html">ViewControls</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertControls.html">InsertControls</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li><li><a href="ViewControls.html">ViewControls</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 08:46:07 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 09:37:08 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
