+++ doc/module-Color.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Class: EditMode</title>
+    <title>JSDoc: Module: Color</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Class: EditMode</h1>
+    <h1 class="page-title">Module: Color</h1>
 
     
 
@@ -28,7 +28,6 @@
 
 <header>
     
-        <h2><span class="attribs"><span class="type-signature"></span></span>EditMode<span class="signature">(neonView, meiFile, zoomHandler)</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -38,10 +37,103 @@
     
         
 
+        
+    
+    </div>
+
     
 
     
-    <h4 class="name" id="EditMode"><span class="type-signature"></span>new EditMode<span class="signature">(neonView, meiFile, zoomHandler)</span><span class="type-signature"></span></h4>
+
+    
+
+    
+
+    
+
+    
+
+    
+        <h3 class="subsection-title">Members</h3>
+
+        
+            
+<h4 class="name" id="~ColorPalette"><span class="type-signature">(inner, constant) </span>ColorPalette<span class="type-signature"></span></h4>
+
+
+
+
+<div class="description">
+    Color palette from Figure 2 (Colors optimized for color-blind
+individuals) from "Points of view: Color blindness" by Bang Wong
+published in Nature Methods volume 8 on 27 May 2011
+https://www.nature.com/articles/nmeth.1618?WT.ec_id=NMETH-201106
+</div>
+
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="Color.js.html">Color.js</a>, <a href="Color.js.html#line61">line 61</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+        
+    
+
+    
+        <h3 class="subsection-title">Methods</h3>
+
+        
+            
+
+    
+
+    
+    <h4 class="name" id=".highlight"><span class="type-signature">(static) </span>highlight<span class="signature">(staff, color)</span><span class="type-signature"></span></h4>
     
 
     
@@ -49,7 +141,7 @@
 
 
 <div class="description">
-    Creates user interface for editing and creates necessary tools.
+    Highlight a staff a certain color.
 </div>
 
 
@@ -85,13 +177,13 @@
 
         <tr>
             
-                <td class="name"><code>neonView</code></td>
+                <td class="name"><code>staff</code></td>
             
 
             <td class="type">
             
                 
-<span class="param-type"><a href="NeonView.html">NeonView</a></span>
+<span class="param-type">SVGSVGElement</span>
 
 
             
@@ -101,14 +193,14 @@
 
             
 
-            <td class="description last">The NeonView parent object.</td>
+            <td class="description last">The staff's SVG element.</td>
         </tr>
 
     
 
         <tr>
             
-                <td class="name"><code>meiFile</code></td>
+                <td class="name"><code>color</code></td>
             
 
             <td class="type">
@@ -124,30 +216,7 @@
 
             
 
-            <td class="description last">The path to the MEi file.</td>
-        </tr>
-
-    
-
-        <tr>
-            
-                <td class="name"><code>zoomHandler</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="module-Zoom-ZoomHandler.html">module:Zoom~ZoomHandler</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The ZoomHandler object.</td>
+            <td class="description last">The color to highlight the staff.</td>
         </tr>
 
     
@@ -188,7 +257,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line17">line 17</a>
+        <a href="Color.js.html">Color.js</a>, <a href="Color.js.html#line26">line 26</a>
     </li></ul></dd>
     
 
@@ -216,33 +285,13 @@
 
 
 
-    
-    </div>
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-        <h3 class="subsection-title">Methods</h3>
-
         
             
 
     
 
     
-    <h4 class="name" id="~init"><span class="type-signature">(inner) </span>init<span class="signature">()</span><span class="type-signature"></span></h4>
+    <h4 class="name" id=".setStaffHighlight"><span class="type-signature">(static) </span>setStaffHighlight<span class="signature">()</span><span class="type-signature"></span></h4>
     
 
     
@@ -250,7 +299,7 @@
 
 
 <div class="description">
-    Initialize handlers and controls and create event listeners.
+    Highlight each staff a different color.
 </div>
 
 
@@ -294,7 +343,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line34">line 34</a>
+        <a href="Color.js.html">Color.js</a>, <a href="Color.js.html#line6">line 6</a>
     </li></ul></dd>
     
 
@@ -328,7 +377,7 @@
     
 
     
-    <h4 class="name" id="~resetListeners"><span class="type-signature">(inner) </span>resetListeners<span class="signature">()</span><span class="type-signature"></span></h4>
+    <h4 class="name" id=".unhighlight"><span class="type-signature">(static) </span>unhighlight<span class="signature">(staff)</span><span class="type-signature"></span></h4>
     
 
     
@@ -336,7 +385,145 @@
 
 
 <div class="description">
-    Reset select event listeners.
+    Remove the highlight from a staff.
+</div>
+
+
+
+
+
+
+
+
+
+    <h5>Parameters:</h5>
+    
+
+<table class="params">
+    <thead>
+    <tr>
+        
+        <th>Name</th>
+        
+
+        <th>Type</th>
+
+        
+
+        
+
+        <th class="last">Description</th>
+    </tr>
+    </thead>
+
+    <tbody>
+    
+
+        <tr>
+            
+                <td class="name"><code>staff</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="param-type">SVGSVGElement</span>
+|
+
+<span class="param-type">string</span>
+
+
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">The staff's SVG element or a JQuery selector.</td>
+        </tr>
+
+    
+    </tbody>
+</table>
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="Color.js.html">Color.js</a>, <a href="Color.js.html#line43">line 43</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        
+            
+
+    
+
+    
+    <h4 class="name" id=".unsetStaffHighlight"><span class="type-signature">(static) </span>unsetStaffHighlight<span class="signature">()</span><span class="type-signature"></span></h4>
+    
+
+    
+
+
+
+<div class="description">
+    Remove the highlight from each staff.
 </div>
 
 
@@ -380,7 +567,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line50">line 50</a>
+        <a href="Color.js.html">Color.js</a>, <a href="Color.js.html#line17">line 17</a>
     </li></ul></dd>
