+++ doc/module-Contents.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Class: EditMode</title>
+    <title>JSDoc: Module: Contents</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Class: EditMode</h1>
+    <h1 class="page-title">Module: Contents</h1>
 
     
 
@@ -28,7 +28,6 @@
 
 <header>
     
-        <h2><span class="attribs"><span class="type-signature"></span></span>EditMode<span class="signature">(neonView, meiFile, zoomHandler)</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -38,184 +37,7 @@
     
         
 
-    
-
-    
-    <h4 class="name" id="EditMode"><span class="type-signature"></span>new EditMode<span class="signature">(neonView, meiFile, zoomHandler)</span><span class="type-signature"></span></h4>
-    
-
-    
-
-
-
-<div class="description">
-    Creates user interface for editing and creates necessary tools.
-</div>
-
-
-
-
-
-
-
-
-
-    <h5>Parameters:</h5>
-    
-
-<table class="params">
-    <thead>
-    <tr>
         
-        <th>Name</th>
-        
-
-        <th>Type</th>
-
-        
-
-        
-
-        <th class="last">Description</th>
-    </tr>
-    </thead>
-
-    <tbody>
-    
-
-        <tr>
-            
-                <td class="name"><code>neonView</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="NeonView.html">NeonView</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The NeonView parent object.</td>
-        </tr>
-
-    
-
-        <tr>
-            
-                <td class="name"><code>meiFile</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type">string</span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The path to the MEi file.</td>
-        </tr>
-
-    
-
-        <tr>
-            
-                <td class="name"><code>zoomHandler</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="module-Zoom-ZoomHandler.html">module:Zoom~ZoomHandler</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The ZoomHandler object.</td>
-        </tr>
-
-    
-    </tbody>
-</table>
-
-
-
-
-
-
-<dl class="details">
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-    <dt class="tag-source">Source:</dt>
-    <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line17">line 17</a>
-    </li></ul></dd>
-    
-
-    
-
-    
-
-    
-</dl>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
     
     </div>
 
@@ -232,34 +54,30 @@
     
 
     
-
-    
-        <h3 class="subsection-title">Methods</h3>
+        <h3 class="subsection-title">Members</h3>
 
         
             
+<h4 class="name" id=".editControlsPanel"><span class="type-signature">(static, constant) </span>editControlsPanel<span class="type-signature"> :string</span></h4>
 
-    
-
-    
-    <h4 class="name" id="~init"><span class="type-signature">(inner) </span>init<span class="signature">()</span><span class="type-signature"></span></h4>
-    
-
-    
 
 
 
 <div class="description">
-    Initialize handlers and controls and create event listeners.
+    Contents of edit panel with buttons.
 </div>
 
 
 
+    <h5>Type:</h5>
+    <ul>
+        <li>
+            
+<span class="param-type">string</span>
 
 
-
-
-
+        </li>
+    </ul>
 
 
 
@@ -294,7 +112,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line34">line 34</a>
+        <a href="Contents.js.html">Contents.js</a>, <a href="Contents.js.html#line87">line 87</a>
     </li></ul></dd>
     
 
@@ -310,42 +128,28 @@
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
         
             
+<h4 class="name" id=".insertControlsPanel"><span class="type-signature">(static, constant) </span>insertControlsPanel<span class="type-signature"> :string</span></h4>
 
-    
-
-    
-    <h4 class="name" id="~resetListeners"><span class="type-signature">(inner) </span>resetListeners<span class="signature">()</span><span class="type-signature"></span></h4>
-    
-
-    
 
 
 
 <div class="description">
-    Reset select event listeners.
+    Structure of insert panel with basic grouping tabs.
 </div>
 
 
 
+    <h5>Type:</h5>
+    <ul>
+        <li>
+            
+<span class="param-type">string</span>
 
 
-
-
-
+        </li>
+    </ul>
 
 
 
@@ -380,7 +184,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="EditMode.js.html">EditMode.js</a>, <a href="EditMode.js.html#line50">line 50</a>
+        <a href="Contents.js.html">Contents.js</a>, <a href="Contents.js.html#line71">line 71</a>
     </li></ul></dd>
     
 
@@ -396,12 +200,72 @@
 
 
 
+        
+            
+<h4 class="name" id=".insertTabHtml"><span class="type-signature">(static, constant) </span>insertTabHtml<span class="type-signature"> :object</span></h4>
+
+
+
+
+<div class="description">
+    HTML for each insert tab (neume, grouping, clef, system, and division).
+</div>
+
+
+
+    <h5>Type:</h5>
+    <ul>
+        <li>
+            
+<span class="param-type">object</span>
+
+
+        </li>
+    </ul>
 
 
 
 
 
+<dl class="details">
 
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="Contents.js.html">Contents.js</a>, <a href="Contents.js.html#line21">line 21</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
 
 
 
@@ -409,6 +273,80 @@
 
 
         
+            
+<h4 class="name" id=".navbarDropdownMenu"><span class="type-signature">(static, constant) </span>navbarDropdownMenu<span class="type-signature"> :string</span></h4>
+
+
+
+
+<div class="description">
+    Contents of navbar menu after switching to edit mode.
+</div>
+
+
+
+    <h5>Type:</h5>
+    <ul>
+        <li>
+            
+<span class="param-type">string</span>
+
+
+        </li>
+    </ul>
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="Contents.js.html">Contents.js</a>, <a href="Contents.js.html#line60">line 60</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+        
+    
+
