+++ doc/module-Zoom-ZoomHandler.html
@@ -29,7 +29,7 @@
 <header>
     
         <h2><span class="attribs"><span class="type-signature"></span></span>
-            <span class="ancestors"><a href="module-ZoomHandler.html">ZoomHandler</a>~</span>ZoomHandler<span class="signature">()</span><span class="type-signature"></span></h2>
+            <span class="ancestors"><a href="module-Zoom.html">Zoom</a>~</span>ZoomHandler<span class="signature">()</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -94,7 +94,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line3">line 3</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line3">line 3</a>
     </li></ul></dd>
     
 
@@ -142,7 +142,7 @@
 
         
             
-<h4 class="name" id="~viewBox"><span class="type-signature">(inner) </span>viewBox<span class="type-signature"></span></h4>
+<h4 class="name" id="~viewBox"><span class="type-signature">(inner) </span>viewBox<span class="type-signature"> :<a href="module-Zoom.ViewBox.html">module:Zoom.ViewBox</a></span></h4>
 
 
 
@@ -152,6 +152,16 @@
 </div>
 
 
+
+    <h5>Type:</h5>
+    <ul>
+        <li>
+            
+<span class="param-type"><a href="module-Zoom.ViewBox.html">module:Zoom.ViewBox</a></span>
+
+
+        </li>
+    </ul>
 
 
 
@@ -186,7 +196,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line10">line 10</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line13">line 13</a>
     </li></ul></dd>
     
 
@@ -266,7 +276,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line62">line 62</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line65">line 65</a>
     </li></ul></dd>
     
 
@@ -352,7 +362,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line16">line 16</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line19">line 19</a>
     </li></ul></dd>
     
 
@@ -438,7 +448,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line50">line 50</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line53">line 53</a>
     </li></ul></dd>
     
 
@@ -596,7 +606,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line41">line 41</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line44">line 44</a>
     </li></ul></dd>
     
 
@@ -731,7 +741,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="ZoomHandler.js.html">ZoomHandler.js</a>, <a href="ZoomHandler.js.html#line29">line 29</a>
+        <a href="Zoom.js.html">Zoom.js</a>, <a href="Zoom.js.html#line32">line 32</a>
     </li></ul></dd>
     
 
@@ -775,13 +785,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.html">ZoomHandler</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertControls.html">InsertControls</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-ZoomHandler.ViewBox.html">ViewBox</a></li><li><a href="module-ZoomHandler-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li><li><a href="ViewControls.html">ViewControls</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul><h3>Global</h3><ul><li><a href="global.html#Navbar">Navbar</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 09:34:25 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 10:56:27 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
