+++ doc/module-Zoom.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Module: Neon</title>
+    <title>JSDoc: Module: Zoom</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Module: Neon</h1>
+    <h1 class="page-title">Module: Zoom</h1>
 
     
 
@@ -49,7 +49,10 @@
         <h3 class="subsection-title">Classes</h3>
 
         <dl>
-            <dt><a href="module-Neon-Neon.html">Neon</a></dt>
+            <dt><a href="module-Zoom.ViewBox.html">ViewBox</a></dt>
+            <dd></dd>
+        
+            <dt><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></dt>
             <dd></dd>
         </dl>
