--- src/ViewControls.js
@@ -1,120 +1,0 @@
-import ColorStaves from "./ColorStaves.js";
-
-/**
- * Set handlers for the display controls.
- * @constructor
- * @param {module:ZoomHandler~ZoomHandler} zoomhandler - An instantiated ZoomHandler.
- */
-function ViewControls (zoomHandler) {
-
-    var color = new ColorStaves();
-    setZoomControls();
-    setOpacityControls();
-    setBackgroundOpacityControls();
-    setSylControls();
-    setHighlightControls();
-
-    $("#displayHeader").on("click", () => {
-        if ($("#displayContents").is(":hidden")) {
-            $("#displayContents").css("display", "");
-        } else {
-            $("#displayContents").css("display", "none");
-        }
-    });
-
-    /**
-     * Set zoom control listeners for button and slider.
-     */
-    function setZoomControls() {
-        $("#reset-zoom").click( function() {
-            $("#zoomOutput").val(100);
-            $("#zoomSlider").val(100);
-            zoomHandler.resetZoomAndPan();
-        });
-
-        $(document).on('input change', '#zoomSlider', function () {
-            zoomHandler.zoomTo($("#zoomOutput").val() / 100.0)
-        })
-    }
-
-    // TODO: Potentially separate opacity into its own js file as well? similar to zoom
-    // Though as it stands its pretty small so it might be fine here
-    /**
-     * Set rendered MEI opacity button and slider listeners.
-     */
-    function setOpacityControls() {
-        $("#reset-opacity").click( function() {
-            // Definition scale is the root element of what is generated by verovio
-            $(".definition-scale").css("opacity", 1);
-
-            $("#opacitySlider").val(100);
-            $("#opacityOutput").val(100);
-        })
-
-        $(document).on('input change', '#opacitySlider', setOpacityFromSlider);
-    }
-
-    /**
-     * Set background image opacity button and slider listeners.
-     */
-    function setBackgroundOpacityControls() {
-        $("#reset-bg-opacity").click( function() {
-            $("#bgimg").css("opacity", 1);
-
-            $("#bgOpacitySlider").val(100);
-            $("#bgOpacityOutput").val(100);
-        })
-
-        $(document).on('input change', '#bgOpacitySlider', function () {
-            $("#bgimg").css("opacity", $("#bgOpacityOutput").val() / 100.0);
-        });
-    }
-
-    /**
-     * Set listener on syllable visibility checkbox.
-     */
-    function setSylControls() {
-        updateSylVisibility();
-        $("#displayText").click(updateSylVisibility);
-    }
-
-    function setOpacityFromSlider() {
-        $(".definition-scale").css("opacity", $("#opacityOutput").val() / 100.0);
-    };
-
-
-    /**
-     * Set listener on staff highlighting checkbox.
-     */
-    function setHighlightControls(view) {
-        updateHighlight();
-        $("#highlightStaves").click(updateHighlight);
-    }
-
-    function updateSylVisibility() {
-        if ($("#displayText").is(":checked")) {
-            $(".syl").css("visibility", "visible");
-        } else {
-            $(".syl").css("visibility", "hidden");
-        }
-    }
-
-    function updateHighlight() {
-        if ($("#highlightStaves").is(":checked")) {
-            color.setColor();
-        } else {
-            color.unsetColor();
-        }
-    }
-
-    ViewControls.prototype.constructor = ViewControls;
-    ViewControls.prototype.setZoomControls = setZoomControls;
-    ViewControls.prototype.setOpacityControls = setOpacityControls;
-    ViewControls.prototype.setBackgroundOpacityControls = setBackgroundOpacityControls;
-    ViewControls.prototype.setSylControls = setSylControls;
-    ViewControls.prototype.setOpacityFromSlider = setOpacityFromSlider;
-    ViewControls.prototype.setHighlightControls = setHighlightControls;
-    ViewControls.prototype.updateHighlight = updateHighlight;
-    ViewControls.prototype.updateSylVisibility = updateSylVisibility;
-}
-export {ViewControls as default};
